Better analysis (inner-bank jumps and calls)
we still need some foo for bankswitches, to be faster and better. This can be done by a reg-profile in this file. I'll do this, but I have to read more about it.

But we also need a new R_ANAL_OP_TYPE, something like R_ANAL_OP_TYPE_JRET or R_ANAL_OP_TYPE_RJMP to follow the program-flow in scenarios like this:

We're in rombank23:
ld a, 7
call 0x2342     ; which is in rombank0(rombank0 is fixed)

in rombank0:
ld 0x2000, a    ; bankswitch
ret             ; bämmm! we return to somewhere else