dwarf: Store DW_AT_comp_dir var in sdb_addrinfo.
This variable points to the default directory to be
searched for source files of the binary. This is used
by linenumber information but is stored in debug_info section.
debug_info by default is DWARF4(which is currently not supported)
and debug_lines is by default DWARF2 so for now if using DWARF4
in your binaries you would have to set this var by hand.