Fix visual assembler
The altered section of code is supposed to update the assembly code displayed
when enter new opcodes. This is done gy giving the hexadecimal representation
of the new binary to the the "pd" command. The corresponding flag is "@x:" and
not "@b:" (the later is used to temporarily set asm.bits).