Fix invalid free after tab-completing pf or te commands
Commands such as "pf." or "te" changed the tmp_argv_heap flag to true,
which means the completion argv is heap allocated and
r_core_free_autocomplete() should free it.

r_core_fgets() assigns a different, non-heap allocated array to argv,
sets argc to 226, and doesn't free the old one or change that flag.
So the next call to r_core_free_autocomplete() tries to free 226 items
out of tmp_argv.

This is fixed by freeing the old tmp_argv before replacing it with
radare_argv