Rewrite the ARCompact analyser code. This started because I noticed that the older code was really not decoding the binary I was working on properly.
Specific things that I noticed were:
* Scrolling backwards scrolled forwards
* Jump targets were shown from apparently random opcodes
* Analysis really didnt find anything
* The anal code was essentially hardcoded for big endian targets (and mine
  was not, so instruction decoding was largely misfiring)

I read the ARCompactISA Programmers Reference guide and implemented decoding
for most instructions.  Specifically, I ensured that all the instructions that
had limm data were decoded with the correct length and that all the jump or
branch targets were correctly calculated.

As I have not previously looked at anal plugins before, I expect I have
messed up a bunch of the assumptions that the anal system makes.

Amongst the issues that I still have is that the analysis will find
instructions (and thus xrefs, functions, etc) in "unaligned" locations.
This might just be an artifact of the fact that not all my data is marked
as data, but it feels like there might be something I have missed.

Still remaining TODO are most of the 32bit general purpose instructions.