core/graph: improve layer assignment algorithm
It should fix some layer problem on some binaries like ls-linux64 and in
the presence of self-loop. The previous algorithm could assign the same
layer to two nodes that are connected with an edge (and it should never
happen, because they have to be on different layers).

This patch uses the topological sort of a graph to assign the layer.