Refactor nz assembler
Beginning to refactor the handmade assembler to incorporate some
concepts of the tab assembler with the hackability of the nz assembler.
The idea is that each instruction that takes arguments is handled by its
own function. Instructions without arguments just resolve to hardcoded
instructions for the lookup table. Special cases are also group 1 and
group 2 instructions

http://www.sandpile.org/x86/opc_grp.htm

A further aim is to remove the tab assetmbler in order to reduce the
number of assemblers that ship with r2.

This squashes 45 commits.

PR amends
---------

* PR fixes for comments by @alvarofe
* PR fixes for comments by @radare