Some improvements for arm pseudo code (#5370)
* Fix logical typo scanning for separator

The test would only accept ] as a separator. This was ok for
expressions like:

  r2 = [pc + 8]

But would strip the terminating ) in:

  if (r1 == 0

* arm/pseudo: fix typo /halt/half

* arm/pseudo: fix spacing before byte/half

it is now consistent with the other instructions and
have a space after the '=' also consistent with vmov/vdif.

* arm/pseudo: fix ldr/str

The following str instruction:

  str r0, [r4, 0xc]

Would produce the broken pseudocode:

  [r4 + 0xc] = r0 + 0xc]

With this change we get the expected:

  [r4 + 0xc] = r0

We need to handle 2/3 parameter instructions
differently for ldr/str. By doing this the
special case checking for [ is no longer needed.

* arm/pseudo: add a couple of XXXXs variants

these were arbitrarily selected based on the sample
i have been looking at.