Replace ${LIB} with lib in bochs Makefile (#5682)
r2 builds currently fail on Bash on Ubuntu on Windows due to the
${NAME} variable being defined in the environment (as the current
hostname), which meant that libbochs.a was being placed in an
unexpected location. This patch fixes the issue for BoUoW by replacing
use of the ${LIB} variable with a set value of 'lib'. It should be
noted that other build scripts (e.g. shlr/bochs/deps.mk) currently
assume that libbochs.a is located precisely at
shlr/bochs/lib/libbochs.a anyway (and as such, it is suspected that
the use of ${LIB} here may have been spurious to begin with).