Fix invalid free() on teardown for io plugins
For example given a plugin defined like this:

RIOPlugin r_io_plugin_foo = {
	.name = "foo",
	...snip...
};

#ifndef CORELIB
RLibStruct radare_plugin = {
	.type = R_LIB_TYPE_IO,
	.data = &r_io_plugin_foo,
	.version = R2_VERSION
};
#endif

We will try to `free(&r_io_plugin_foo)`, which was not allocated on the
heap. Static io plugins are OTOH allocated like this, so we need to do
the same for dynamically loaded io plugins. However, this is not the same
across the different plugin types, so we need to special-case io plugins
here.