Debian patches (#5866)
* Fix esil manpage

Each manual page should start with a "NAME" section, which lists the
name and a brief description of the page separated by "\-". The "NAME"
section is parsed by lexgrog and used to generate a database that's
queried by commands like apropos and whatis. Without this change
lexgrog fails to parse the NAME section of this manual page.

This was found by Debian's lintian tool (manpage-has-bad-whatis-entry).

* Fix empty input line in manpages

Found-By: lintian as manpage-has-errors-from-man

* Fix rahash2 manpage

.El (end list) is only valid after .Bl (begin list),
but no list was started in this section.

Found-By: lintian

* Avoid build break if no static bin plugins are defined

If one disables all static bin plugins, build currently fails,
since R_BIN_STATIC_PLUGINS/R_BIN_XTR_STATIC_PLUGINS are not
defined. This fixes the build for these kind of setups.

(In Debian does not built the mach code, since the
Apple Public Source License is not compatible with
the Debian Free Software Guidelines. Without the
mach code there is not a single binxtr plugin)

* Fix implicit declaration of function 'linux_generate_corefile' on i386

Noticed by Debian build farm.

* Add support for SOURCE_DATE_EPOCH

SOURCE_DATE_EPOCH can be used to make builds
reproducible [0].

Note: radare suffers from a second problem making the
build unreproducible: It captures the full build path.
This is *not* covered by this change.

[0] https://reproducible-builds.org/specs/source-date-epoch/