Fix #2401, #2402, #6133 - p8/pB/px with negative len
p8 (#2401) and pB (#2402) were broken, and px wasn't cleaning up
correctly after changing the block size (#6133).

The fix for the last one involved deferring block size cleanup to the
r_core_block_size() call at the end of the function, and moving that
call *after* seeking to the original position instead of before (which
is what was making it leave data with the wrong offset in core->block)

After simplifying that, the fix to the p8 and pB issue is just calling
r_core_block_size() to reread the block like px does.