Added help message when izz after -nn
When opening a binary with -nn internal structures are not initialized being
one of them RBinFile that is need to retrieve the info.

However, when opening an image for example RBinFile exist but we do not have a
corresponding plugin or binary format to say so we are able to get strings.
This will be need a great refactor to handle correctly and furthermore
something I did not like is that baddr is only accesible through RBinObject and
I think it should be within RBin at least.

This was working before cause rabin2 was being called but without -B as
parameter. I have resorted for eprintf message so the user can configure as his
need.