Add &nbsp; to html output and fix <br />
e scr.html=true doesn't print spaces as &nbsp; and thus it looks messy because the maximum spaces sequence allowed in html is 1. 
Therefore we should change every space to &nbsp;. 

In addition, fix a problem with <br /> printing, \n (a new line) was printed instead and it doesn't consider as line-break in some editors. Therefore now it avoids printing new line and prints <br /> instead.