Fix generated ESIL for AVR flags.
A typo of `__generic_sub_update_flags_rr` meant that the generated
code contained the immediate as a register, resulting in the flags
not being set correctly.

After switching this to `_rk` the format of the immediate values were
still incorrect because they were being output as hex without the `0x`
prefix. This was changed to output decimal instead, as that matches the
format of the result value when doing operations like `cpi`. For
example:

    cpi r26, 0x2e
    46,r26,-,r26,0x08,&,!,46,0x08,&...

The format for the `__generic_add_update_flags` was also fixed, but
as there is currently no `_rk` version it doesn't affect anything.