fix segfault in vsnprintf on 32 bit systems examining a PPC ELF (#8075)
* fix segfault in vsnprintf on 32 bit systems 

On a 32-bit x86 computer, examining a PowerPC ELF with the "aa" command caused a segfault in vsnprintf. 

The %llx  (PFMT64x) format attempts to get the next 64 bits from the stack. This will consume the cmask32 and ARG(0) values. The following "%s" will cause a segfault if the next 32 bits is not pointing to readable memory.

I doubt this is the best fix. Maybe libr/include/r_types.h should be changed? You guys know better.

* cast instead of duplicating code

* fix white space