Fix creation of invalid function refs when saving/loading a project.
This resolves radare/radare2#7774.

When saving a project, a dump of the output from the "afl*" command is
used to store all the function references in the project rc file. The
handler for "afl*" generates many afx[cCd] commands to add function
references. However, instead of using the actual reference from
address, the handler uses the function root address as the from
address for the afx[cCd] command. After loading a project, the
execution of the commands in the project rc file causes the invalid
function xrefs to be created, as described in radare/radare2#7774.

The fix does two things. First it changes the way the afx[cCd]
commands are generated by the "afl*" handler so it uses the actual
reference from address instead of the function root address for the
afx[cCd] from address. Second, it changes the afx[cCds] command
handler to add function references even when the reference from
address is not necessarily a function root. I'm not sure why the
original code required the from address to be a function root.

This fixes the output of commands axt and afx after loading a
project. They now match the command output prior to saving a project.