MZ: Fix header check for minimal headers
MZ header length is measured in 16-byte paragraphs. A minimum header is
0x1B bytes which means the header is padded to 0x20 bytes. MZ header
check was checking that the entrypoint address was > 0x40 which assumes
a certain number of relocations or other header paragraphs.
Additionally, the CS was being masked off which applied the above check
even to entrypoints far away from the header (which is entirely in
CS==0).