dreg: Fix printing of 16-bit x86 segment registers
'ar seg' ends up calling r_debug_reg_list() with size=0.  This should
disable filtering of matches by size per the inner loop iterating over
the registers.  An earlier check to ensure the requested size is valid
for this register set fails to allow for size=0 and will force it to
size=32 instead.  When in 16-bit mode, that will cause no registers to
be printed.