Fix tmpseek when execution times is in the command
Before applying this patch, running `times cmd @ tmpoff` will seek to
*tmpoff* and run *cmd* the first time, restore to the original offset,
then run *cmd* at the original offset for the rest of the time. This
will break the `wxs` command which changes the offset after each run.

In this patch, when tmpseek is used in `[rep] cmd @ tmpoff`:
1. save core->offset, seek to tmpoff
2. run the command rep times, do not restore the offset after a run
3. then restore core->offset even if cmd changes offset

Test:
- s addr @ tmpoff: this should not change the offset
- rep wxs bytes @ tmpoff: this should write bytes *rep* times
  starting at *tmpoff*, then the offset is restored

More tests are needed