Consolidate thumb BL and BLX, simplify, bugfix
Encoding for BL and BLX with immediate offset is identical, only
difference is the opcode in the second half of the instruction pair.

Use r_num_math instead of getnum, as the latter does not work correctly
for large arguments (>= 0x80000000).

Simplify logic for 16 bit aligned origin addresses. In case the origin
is not 32 bit aligned copy bit[1] to the target address (which is masked
by the decoder) and calculate the offset based on the modified address.

Using the same implementaton also fixes the missing offset handling in BL.

Fix for #9319