arm: Don't set bits hint in anop32
A hint was set in anop32 when seeing a blx instruction. This was causing
glitches while disassembling raw arm32 code: suddenly asm.bits would
jump to 16, and it was unclear how to change it back: `e asm.bits=32`
would not stick.

Fix proposed by pancake on IRC.

The removed code was introduced in 2ec6722bfecfe02c41089339fa2940a1a9194deb.