io: split functions in regular and _batch one instead of do_skyline
This way it is easier to understand the difference between the two
things. A parameter hides the complexity in a simple true/false, but the
two functions (regular and _batch) are *very* different and require
special attention by the programmer since he needs to manually call the
update_skyline function to maintain the internal consistency of IO.

In general the regular versions should be used, unless for performance
reasons we think we can avoid the update of the skyline on every call.