Fix r_hex_str_is_valid to handle space like r_hex_str2bin
r_hex_str_is_valid is called only once from r_reg_arena_set_bytes,
to find the size of the hexadecimal string. However, it stop
at the first space, while the method used for the conversion
r_hex_str2bin is skipping space and continue to convert.

So with a long enough string passed to drw and arw, it is possible
to crash r2, as reported in https://github.com/radare/radare2/issues/11407