Remove global for addr since that's uneeded
Since we are using the name addr all around, lgtm.com complain
about it (and likely other linters). Looking at the issue, it seems
that we do not need a global variable, since addr is reset each time
at the start of r_core_visual_anal, and seems to be unused by others
functions.