format/elf: fix is_in_pphdr function (fix #11377) (#11683)
* format/elf: fix is_in_pphdr function

That function should check if an address is in the file "side" of the
phdr, so it should check for filesz and not memsz

* format/elf: consider filesz even when looking at virtual space

A virtual address that is included in p_memsz but not in p_filesz, means
that it's an address that has no counterpart in the physical file, does
it makes no sense to look at memsz when converting an address from
virtual to physical.

* bin/demangle: demangle names that end with @@CXXABI

After last commit, we get correct mangled names that, according to other
tools, include also the @@CXXABI part. This patch removes the @@CXXABI
suffix because the cplus_demangle_v3 function is not able to deal with
it.