Support PDB ages higher than 15 (#11805)
`GUIDSTR_LEN` is currently set to 34, which gives 32 characters for the PDB GUID, 1 character for the PDB age (which is tacked on the end), and the null terminator. However, PDB ages are stored as a uint32, and can thus have a maximum value of `0xFFFFFFFF`, which requires up to 8 characters. PDB ages over 15 can easily be reached by automatic build processes using incremental linking.