Fix another assert when printing blocks (#12173)
* Do not NULLIFY the result if the list is empty when getting boundaries

An empty list is different than a NULL return value. The caller has to
call r_list_free anyway on the returned list, so there's no real
advantage in returning NULL, it just makes the caller's life harder
because it cannot understand if NULL is returned because the list was
empty (e.g. there are no IO maps, etc.) or because there were some
errors.

* Check for NULL and exit if so, that's an error.

* Check r_core_get_boundaries_prot return value for NULL