types: allow an offset to be set for the struct itself
Not all pointers to structs start at the beginning, thanks Microsoft
with your implementation of list entries. Therefore this commit allows
you to tweak the 'entry point' to the struct when using the ta command.
NOTE: Annoyingly this file is formatted badly so there are lots of
additions due to linting!