Move #endif (HAVE_PTRACE) after functions which use it
When built with --disable-debugger, HAVE_PTRACE is not defined. The
function r_io_ptrace_func() is outside of an "#if HAVE_PTRACE" block and
uses ptrace_* functions. This causes a build failure because the
functions are only defined under HAVE_PTRACE.

Move the adjacent "#endif" for a HAVE_PTRACE guard to include this
function too.