anal_riscv: Implement get_reg_profile
Add `get_reg_profile` call for RISC-V. In this profile, include the
integer and floating point registers. Use the ABI register naming
instead of the architectural one as this was chosen in the disassembler.

This is straightforward as the user-level ISA only specifies these
registers. In RV-I there are no flags, and only a few read-only
counter CSRs are accessible from user mode which (I think) don't need to
be in this structure. In RV-F/D there is one 32-bit CSR containing a few
flags.

This addition is not quite enough to get analysis or debugging working,
but can be a step towards that, and prevents warnings about being unable
to import the register profile from isil during launch on RISC-V.

Also update bits to `32|64` instead of `16|32`—there is no 16-bit
RISC-V.