Call remove_offsetmap only if the flagItem is not new ##flags
* Call remove_offsetmap only if the flagItem is not new

It does not have sense to call that function on new FlagItems, it just
slows down things for nothing, since item->offset has not been set yet.

* force and is_new must be two separate things

When setting a flag (which may be even a set of some properties of an
existing flagitem) we want to force the update of all the RFlag
properties, but remove_offsetmap should not be called when the flagitem
is new. So we need to track the two properties separately.