Add info that r2pm -i is also used for pkg update (#14998)
It is confusing that `r2pm update` only updates the database
and the packages are not updated. Information about
how to update a specific package is missing.
When the user searches for information how to update a particular package,
e.g. using `r2pm -h`, the help provides information only how to
install a package.