[WASM] Remove incorrect early exit in locals processing
In `consume_locals_r` the following early return check is made:

   if (!(cur + (count * 7) <= max)) { // worst case 7 bytes
           return 0;
   }

However, this rejects valid WASM code. Fore example, consider
the following:

   (module
      (func (export "foo") (result i32)
         (local i32 i64)
         (return (i32.const 0)))
   )

For this case there will be two entries in the locals declaration
vector. One for the `i32` and one for the `i64`. Each entry in the
vector will contain a `u32` for a count representing the number of
locals of that type and a byte representing the value type [1]. Per
the WASM spec a `u32` is at most `ceil(32/7) == 5` bytes [2].

For this particular example, each entry is two bytes -- one for the
count and one for the type. However, the function body is small so
the `cur + (count * 7)` check fails. This patch fixes the problem
by removing the check and letting the locals processing code properly
handle bounds conditions (which it already does).

[1] https://webassembly.github.io/spec/core/binary/modules.html#code-section
[2] https://webassembly.github.io/spec/core/binary/values.html#binary-int