[WASM] Fix incorrect global init expr parsing
Consider the following WASM:

   (module
      (global $x (mut i32) (i32.const 0))
   )

The while loop that processes this code in `consume_init_expr_r`
has two conditions that may cause an exit (1) the current buffer
location is past the "max" mark or (2) the current byte from the
buffer is equal to the `end` opcode (0x0B):

   while (r_buf_tell (b) <= max && r_buf_read8 (b) != eoc) {
           res++;
   }
   if (r_buf_read8 (b) != eoc) {
           return 0;
   }

However, for the (2) condition the above code is incorrect because
`r_buf_read8` bumps the buffer cursor. Thus when it is checked again
in the follow if statement, the cursor is already past the `end`
opcode.

This patch fixes this problem by saving the last byte read in the
loop and checking it in the following if statement.