Fix progressbar rounding error
An extra char was added when the percentage was 100, caused by
what I believe is a rounding error. An int is good enough here.

By removing the double it fixes the min and max macros beeing used
with a mix between integer and double, and the double beeing used
in printf as a decimal value.